chore(design-system): match studio font stack
Align design system with Inter, Manrope, and Source Code Pro so documented components match Studio, www, and docs.
