+++ apps/design-system/app/layout.tsx
@@ -3,11 +3,11 @@ import '@/styles/globals.css'
 
 import type { Metadata, Viewport } from 'next'
 
-import { customFont, sourceCodePro } from './fonts'
 import { Providers } from './Providers'
 import { Toaster } from './toaster'
+import { inter, manrope, sourceCodePro } from '@/lib/fonts'
 
-const className = `${customFont.variable} ${sourceCodePro.variable}`
+const className = `${inter.variable} ${manrope.variable} ${sourceCodePro.variable}`
 
 const BASE_PATH = process.env.NEXT_PUBLIC_BASE_PATH || '/design-system'
 
@@ -126,7 +126,7 @@ export default async function Layout({ children }: RootLayoutProps) {
         {/* [Danny]: This has to be an inline style tag here and not a separate component due to next/font */}
         <style
           dangerouslySetInnerHTML={{
-            __html: `:root{--font-custom:${customFont.style.fontFamily};--font-source-code-pro:${sourceCodePro.style.fontFamily};}`,
+            __html: `:root{--font-sans:${inter.style.fontFamily};--font-heading:${manrope.style.fontFamily};--font-source-code-pro:${sourceCodePro.style.fontFamily};}`,
           }}
         />
       </head>
